Move distutils command in setup.py
