--- maintenance/__init__.py
@@ -1,3 +1,0 @@
-from .i18n import ExtractMessagesCommand
-
-__all__ = ["ExtractMessagesCommand"]
