+++ setup.py
@@ -1,17 +1,48 @@
 #!/usr/bin/env python3
 
+import distutils.cmd
 import sys
 from os.path import dirname, join
 
 from setuptools import find_packages, setup
 
-import maintenance
 from cjwmodule import __version__
 
 # We use the README as the long_description
 readme = open(join(dirname(__file__), "README.md")).read()
 
 needs_pytest = {"pytest", "test", "ptr"}.intersection(sys.argv)
+
+
+class ExtractMessagesCommand(distutils.cmd.Command):
+    """A custom command to run i18n-related stuff."""
+
+    description = "extract i18n messages or check if they need extraction"
+    user_options = [
+        # The format is (long option, short option, description).
+        (
+            "check",
+            None,
+            "Check if all messages have been extracted without modifying catalogs",
+        ),
+    ]
+
+    def initialize_options(self):
+        """Set default values for options."""
+        # Each user option must be listed here with their default value.
+        self.check = False
+
+    def finalize_options(self):
+        """Post-process options."""
+
+    def run(self):
+        from maintenance.i18n import check, extract
+
+        """Run command."""
+        if self.check:
+            check()
+        else:
+            extract()
 
 
 setup(
@@ -35,5 +66,5 @@ setup(
         "tests": ["pytest~=5.3.0", "pytest-asyncio~=0.10.0"],
         "maintenance": ["babel~=2.8.0"],
     },
-    cmdclass={"extract_messages": maintenance.ExtractMessagesCommand,},
+    cmdclass={"extract_messages": ExtractMessagesCommand},
 )
